Rename std package to instances
For #1135.

So far I haven't renamed the `std` within `kernel`. I'm happy to make
those changes as well, but I wanted to check with @non and @johnynek.

I also didn't rename all of the `catsStdX` instances. Those names seemed
fine to me, and some of them have names like `catsStdInstancesForList`,
which doesn't translate well if you switch `Std` to `Instances`.
