+++ core/src/main/scala/cats/implicits/package.scala
@@ -1,3 +1,3 @@
 package cats
 
-object implicits extends syntax.AllSyntax with std.AllInstances
+object implicits extends syntax.AllSyntax with instances.AllInstances
