+++ core/src/main/scala/cats/std/bigInt.scala
@@ -1,5 +1,5 @@
 package cats
-package std
+package instances
 
 trait BigIntInstances extends cats.kernel.std.BigIntInstances {
   implicit val catsStdShowForBigInt: Show[BigInt] =
