+++ core/src/main/scala/cats/std/map.scala
@@ -1,5 +1,5 @@
 package cats
-package std
+package instances
 
 trait MapInstances extends cats.kernel.std.MapInstances {
