+++ tests/src/test/scala/cats/tests/SyntaxTests.scala
@@ -1,7 +1,7 @@
 package cats
 package tests
 
-import cats.std.AllInstances
+import cats.instances.AllInstances
 import cats.syntax.AllSyntax
 import cats.functor.{Invariant, Contravariant}
